Add offsetable
