+++ lib/offsetable.rb
@@ -1,0 +1,9 @@
+module Offsetable
+  def offsets(date)
+    offsets = (date.to_i**2).digits.first(4)
+    { a: offsets[3],
+      b: offsets[2],
+      c: offsets[1],
+      d: offsets[0] }
+  end
+end
